+++ src/commands/privacy.js
@@ -1,10 +1,10 @@
 import { InteractionResponseType, MessageFlags } from 'discord-api-types/payloads/v9';
 import { createEmbed } from '../utils/embed.js';
-import Privacy from '../utils/privacy.js';
+import Privacy from '../utils/strings/privacy.js';
 
 export default {
     name: 'privacy',
-    description: 'View the privacy policy for DNS over Discord',
+    description: 'View the Privacy Policy for DNS over Discord',
     execute: async ({ response }) => response({
         type: InteractionResponseType.ChannelMessageWithSource,
         data: {
